requirements.txt
cffi==1.15.1
cryptography==41.0.1
gitdb==4.0.10
GitPython==3.1.31
pycparser==2.21
smmap==5.0.0
